Create a modular structure for citation table with the following organization: - Core/ - Base components and utilities   - Formatters/ - Citation formatting logic   - Renderers/ - Table rendering components   - Stylesheets/ - CSS styling   - Elements/ - UI elements like buttons and messages   - FormValidators/ - Form validation logic
