+++ classes/components/forms/CitationStyles/Core/Elements/Buttons.php
@@ -1,0 +1,12 @@
+<?php namespace PKP\components\forms\CitationStyles\Core\Elements;
+
+class Buttons {
+    public static function getFormSaveButton() {
+        return '</table>
+                    <button type="submit" class="save-btn-citations">
+                        ' . __('plugins.generic.jatsParser.citationtable.savebuttontext') . '
+                    </button>
+                </form>
+                </div>';
+    }
+}
